REST API/pyproject.toml
[tool.poetry]
name = "hw13"
version = "0.1.0"
description = ""
authors = ["MatsiukRoman <r.matsuk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.35"
fastapi = "^0.115.0"
python-jose = "^3.3.0"
pydantic = "^2.9.2"
fastapi-mail = "^1.5.0"
python-dotenv = "^1.1.1"
pydantic-settings = "^2.10.1"
slowapi = "^0.1.9"
cloudinary = "^1.44.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
